Removing a SideShow application now also removes its associated content. Content updates are now supported.
git-svn-id: http://lufa-lib.googlecode.com/svn@352 d5102386-fcda-11dd-9fdb-3debd5008f28

+++ Demos/Sideshow/Descriptors.h
@@ -56,7 +56,7 @@
 		{
 			USB_Descriptor_Header_t                Header;
 			
-			int16_t                                Signature[7];
+			int                                    Signature[7];
 			uint16_t                               VendorCode;
 		} USB_OSDescriptor_t;
